feat(api): Dış API entegrasyon altyapısı, DTO düzenlemeleri ve hata yönetimi eklendi (#4)
* refactor(backend): Strateji Tasarım Deseni ve Controller düzenlemesi

- Veri erişim mantığını esnetmek amacıyla IBudgetDataProvider, DbBudgetDataProvider ve ExternalApiBudgetDataProvider sınıfları eklendi.
- Veri kaynaklarının(Database veya Dış API) tip güvenliğiyle seçilebilmesi için DataSourceType enumu eklendi.
- DTO nesneleri okunan veri yapısıyla uyumlu hale getirildi.
- Controller üzerine eklenmiş olan hesaplama iş mantığı BudgetPredictionManager servisine aktarıldı.

* feat: dış API hata yönetimi ve katmanlı mimarı iyileştirmeleri

- Dış servis hataları için ExternalApiException ve ExternalApiErrorCode enum yapısına geçildi.

- Controller üzerindeki hata yakalama blokları güncellenerek 503, 404 ve 400 yanıtları düzenlendi.

- Program.cs içerisine HttpClient ve Provider servis kayıtları eklendi.

- Kullanılmayan kütüphane bağımlılıkları (using satırları) temizlendi.

---------

+++ BudgetForecastingAPI/DTOs/BudgetPredictionResponseDTO.cs
@@ -6,7 +6,7 @@ namespace BudgetForecastingAPI.DTOs
         public int TahminYili { get; set; }
         public decimal OncekiYilHarcamasi { get; set; }
         public decimal TahminiButce { get; set; }
-        public decimal YuzdeDegisim { get; set; }
+        public double YuzdeDegisim { get; set; }
         public string Aciklama { get; set; } = string.Empty;
     }
 }
